Rework logic in DeleteNodes()
This reworks the logic in DeleteNodes([]nodes) to:

- annotate a node for deletion
- drop the replica count by 1 immediately
- fail fast on any error during those steps

This is different from the current behaviour which first annotated all
the nodes then dropped the replica count by len(nodes). The goal here
is to not leave machines and the respective machine set replica count
in an indeterminate state should any error occur.